# Api/api/docs/swagger.yaml
definitions:
  model.Brand:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.Brands:
    properties:
      brands:
        items:
          $ref: '#/definitions/model.Brand'
        type: array
    type: object
  model.CategiryReq:
    properties:
      name:
        type: string
    type: object
  model.Categories:
    properties:
      categories:
        items:
          $ref: '#/definitions/model.Category'
        type: array
    type: object
  model.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.CreateBrandReq:
    properties:
      name:
        type: string
    type: object
...
  model.Error:
    properties:
      message:
        type: string
    type: object
...
  model.Post:
    properties:
      author_id:
        type: string
      body:
        type: string
      brand_id:
        type: string
      category_id:
        type: string
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      gen:
        type: string
      id:
        type: string
      price:
        type: string
      product_type:
        type: string
      rating:
        type: string
      size:
        items:
          type: string
        type: array
      stars:
        maxLength: 5
        minLength: 0
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - brand_id
    - category_id
    - gen
    - stars
    type: object
  model.Posts:
    properties:
      Posts:
        items:
          $ref: '#/definitions/model.Post'
        type: array
    type: object
...
  model.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model.Error'
    type: object
...
info:
  contact: {}
paths:
  /v1/category:
    post:
      consumes:
      - application/json
      parameters:
      - description: CreateCategory
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategiryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: CreateCategory
      tags:
      - category
  /v1/category/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting category by ID
      parameters:
      - description: Category_ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: Delete Category By ID
      tags:
      - category
    get:
      consumes:
      - application/json
      description: This API for getting posts by category ID
      parameters:
      - description: Category_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Posts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: Get All posts by Category
      tags:
      - category
  /v1/category/getAll:
    get:
      consumes:
      - application/json
      description: This API for getting all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Categories'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      summary: GetAllCategories
      tags:
      - category
  /v1/category/getById/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting Category by ID
      parameters:
      - description: Category_ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: GetCategory
      tags:
      - category
...
  /v1/post/brand:
    post:
      consumes:
      - application/json
      description: This API for creating new brand
      parameters:
      - description: CreateBrand
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/model.CreateBrandReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      tags:
      - brand
  /v1/post/brand/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting brand
      parameters:
      - description: Brand_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: DeleteBrand
      tags:
      - brand
  /v1/post/brand/getAll:
    get:
      consumes:
      - application/json
      description: This API for getting all brands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Brands'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      summary: Get All Brands
      tags:
      - brand
  /v1/post/brand/getByid/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting brand by id
      parameters:
      - description: Brand_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: GetBrandById
      tags:
      - brand
  /v1/post/brand/getPost/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting post bu brand
      parameters:
      - description: Brand_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Posts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: GetPostByBrand
      tags:
      - brand
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"